sep cli app from app data service
